requirements.txt
scapy>=2.5.0
netifaces>=0.11.0
volatility3>=2.4.1
requests>=2.31.0
paramiko>=3.3.1
cryptography>=41.0.0
python-nmap>=0.7.1
pyOpenSSL>=23.2.0
beautifulsoup4>=4.12.2
pillow>=10.0.0
reportlab>=4.0.4
